refactor: move `readPackageJson` and `writePackageJson` in their own helper
