+++ lib/cli/src/detect.test.ts
@@ -1,12 +1,16 @@
 import fs from 'fs';
 
-import { getBowerJson, getPackageJson } from './helpers';
+import { getBowerJson } from './helpers';
 import { isStorybookInstalled, detectFrameworkPreset, detect, detectLanguage } from './detect';
 import { ProjectType, SUPPORTED_FRAMEWORKS, SupportedLanguage } from './project_types';
+import { readPackageJson } from './js-package-manager';
 
 jest.mock('./helpers', () => ({
   getBowerJson: jest.fn(),
-  getPackageJson: jest.fn(),
+}));
+
+jest.mock('./js-package-manager', () => ({
+  readPackageJson: jest.fn(),
 }));
 
 jest.mock('fs', () => ({
@@ -209,18 +213,18 @@ const MOCK_FRAMEWORK_FILES = [
 
 describe('Detect', () => {
   it(`should return type HTML if html option is passed`, () => {
-    (getPackageJson as jest.Mock).mockImplementation(() => true);
+    (readPackageJson as jest.Mock).mockImplementation(() => true);
     expect(detect({ html: true })).toBe(ProjectType.HTML);
   });
 
   it(`should return type UNDETECTED if neither packageJson or bowerJson exist`, () => {
-    (getPackageJson as jest.Mock).mockImplementation(() => false);
+    (readPackageJson as jest.Mock).mockImplementation(() => false);
     (getBowerJson as jest.Mock).mockImplementation(() => false);
     expect(detect()).toBe(ProjectType.UNDETECTED);
   });
 
   it(`should return language typescript if the dependency is present`, () => {
-    (getPackageJson as jest.Mock).mockImplementation(() => ({
+    (readPackageJson as jest.Mock).mockImplementation(() => ({
       dependencies: {
         typescript: '1.0.0',
       },
@@ -229,7 +233,7 @@ describe('Detect', () => {
   });
 
   it(`should return language javascript by default`, () => {
-    (getPackageJson as jest.Mock).mockImplementation(() => true);
+    (readPackageJson as jest.Mock).mockImplementation(() => true);
     expect(detectLanguage()).toBe(SupportedLanguage.JAVASCRIPT);
   });
