+++ lib/cli/src/generators/MARIONETTE/index.ts
@@ -1,7 +1,8 @@
 import fse from 'fs-extra';
 import path from 'path';
-import { writePackageJson, getBabelDependencies } from '../../helpers';
+import { getBabelDependencies } from '../../helpers';
 import { Generator } from '../Generator';
+import { writePackageJson } from '../../js-package-manager';
 
 const generator: Generator = async (packageManager, npmOptions) => {
   const storybookVersion = await packageManager.getVersion('@storybook/marionette');
