+++ lib/cli/src/generators/REACT_NATIVE/index.ts
@@ -1,9 +1,9 @@
 import shell from 'shelljs';
 import chalk from 'chalk';
-import { writePackageJson, paddedLog, getBabelDependencies, copyTemplate } from '../../helpers';
+import { paddedLog, getBabelDependencies, copyTemplate } from '../../helpers';
 import { NpmOptions } from '../../NpmOptions';
 import { GeneratorOptions } from '../Generator';
-import { JsPackageManager } from '../../js-package-manager';
+import { JsPackageManager, writePackageJson } from '../../js-package-manager';
 
 export default async (
   packageManager: JsPackageManager,
