+++ lib/cli/src/generators/SFC_VUE/index.ts
@@ -1,6 +1,7 @@
-import { writePackageJson, getBabelDependencies, copyTemplate } from '../../helpers';
+import { getBabelDependencies, copyTemplate } from '../../helpers';
 import { StoryFormat } from '../../project_types';
 import { Generator } from '../Generator';
+import { writePackageJson } from '../../js-package-manager';
 
 const generator: Generator = async (packageManager, npmOptions, { storyFormat }) => {
   const packages = [
