+++ lib/cli/src/generators/VUE/index.ts
@@ -1,11 +1,11 @@
 import {
-  writePackageJson,
   getBabelDependencies,
   addToDevDependenciesIfNotPresent,
   copyTemplate,
 } from '../../helpers';
 import { StoryFormat } from '../../project_types';
 import { Generator } from '../Generator';
+import { writePackageJson } from '../../js-package-manager';
 
 const generator: Generator = async (packageManager, npmOptions, { storyFormat }) => {
   const packages = [
