+++ lib/cli/src/js-package-manager/index.ts
@@ -1,3 +1,4 @@
 export * from './JsPackageManagerFactory';
 export * from './JsPackageManager';
 export * from './PackageJson';
+export * from './PackageJsonHelper';
